% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypermutR.R
\name{ancestor_processing}
\alias{ancestor_processing}
\title{Internal function to resolve the ancestor argument of remove_hypermut}
\usage{
ancestor_processing(ancestor, dat)
}
\arguments{
\item{ancestor}{Either 'consensus' to indicate that the consensus sequences must be computed, or 'first' to indicate that the first sequence in the dataset should be considered to be the ancestral sequence, or the ancestral sequence itself.}

\item{dat}{The sequence data}
}
\description{
Internal function to resolve the ancestor argument of remove_hypermut
}
