% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{convert_alignment_to_matrix}
\alias{convert_alignment_to_matrix}
\title{Converts an alignment to a matrix}
\usage{
convert_alignment_to_matrix(dat)
}
\arguments{
\item{dat}{DNAStringSet in which each sequence is the same length}
}
\value{
A character matrix
}
\description{
Converts an alignment to a matrix
}
