% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{deduplicate_seqs}
\alias{deduplicate_seqs}
\title{Deduplicates sequence data}
\usage{
deduplicate_seqs(dat)
}
\arguments{
\item{dat}{The sequence data (DNAStringSet)}
}
\description{
Deduplicates sequence data
}
