% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hd_seqs}
\alias{hd_seqs}
\title{High diversity sequence sample}
\format{A SeqFastadna object from library seqinr}
\source{
Simulated based on samples from the Molecular Clock Project
}
\usage{
hd_seqs
}
\description{
High diversity sequence sample
}
\keyword{datasets}
