% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ld_seqs}
\alias{ld_seqs}
\title{Low diversity sequence sample}
\format{A SeqFastadna object from library seqinr}
\source{
Simulated based on samples from the Molecular Clock Project
}
\usage{
ld_seqs
}
\description{
Low diversity sequence sample
}
\keyword{datasets}
