% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{make_list_SeqFastadna}
\alias{make_list_SeqFastadna}
\title{Converts a vector of character strings to the read.fasta format}
\usage{
make_list_SeqFastadna(dat)
}
\arguments{
\item{dat}{Named vector of character strings}
}
\description{
Converts a vector of character strings to the read.fasta format
}
