% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_utils.R
\name{restart_r}
\alias{restart_r}
\title{Restart R}
\usage{
restart_r(status = 0, debug = TRUE)
}
\description{
Useful for developing with the vim-R plugin
From https://stackoverflow.com/questions/6313079/quit-and-restart-a-clean-r-session-from-within-r
}
