% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_hyper}
\alias{sim_hyper}
\title{Simulates hypermutation in sequence datasets}
\usage{
sim_hyper(dat, n1, n2, n3, seed = NULL, verbose = FALSE)
}
\arguments{
\item{dat}{An alignment as produced by seqinr::read.fasta}

\item{n1}{If less than one, the proportion of sequences to mutate, else
the number of sequences to mutate.}

\item{n2}{If less than one, the proportion of potential hypermutation
positions to mutate, else the number of potential hypermutation positions to
mutate. The value 'all' is allowed.}

\item{n3}{If less than one, the proportion of potential control
positions to mutate, else the number of potential control positions to
mutate. The value 'all' is allowed.}

\item{seed}{A seed value to use for the simulations.}
}
\description{
Simulates hypermutation in sequence datasets
}
