% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{deduplicate_seqs}
\alias{deduplicate_seqs}
\title{Deduplicates sequence data}
\usage{
deduplicate_seqs(dat)
}
\arguments{
\item{dat}{The sequence data (SeqFastadna)}
}
\value{
A list in which each unique sequence has an entry consisting of:
\enumerate{
  \item the_seq: The sequence as a character string.
  \item dup_names: A character vector of the headers of all the
   sequences that had that exact sequence.
}
}
\description{
Deduplicates sequence data
}
\details{
The dataset is converted to a vector of character strings and the unique
sequences are selected with the unique function. Looping over the unique
sequences, a list is constructed in which each element corresponds to a
unique sequence. Each element is also a list with the elements the_seq
containing the actual sequences and dup_names, a vector of character strings
listing the names of all sequences that matches the unique sequences stored
in the_seq.
}
\examples{
deduplicate_seqs(ld_seqs)
deduplicate_seqs(c('aaa', 'aaa', 'aab'))

}
