% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{deduplicate_seqs}
\alias{deduplicate_seqs}
\title{Deduplicates sequence data}
\usage{
deduplicate_seqs(dat)
}
\arguments{
\item{dat}{The sequence data (SeqFastadna)}
}
\value{
A list in which each unique sequence has an entry consisting of:
1. the_seq containing the sequence as a character string.
2. dup_names which is a character vector of the headers of all the
   sequences that had that exact sequence.
}
\description{
Deduplicates sequence data
}
