% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{make_list_SeqFastadna}
\alias{make_list_SeqFastadna}
\title{Formats character vector as seqinr::read.fasta}
\usage{
make_list_SeqFastadna(dat)
}
\arguments{
\item{dat}{Named vector of character strings}
}
\description{
seqinr's read.fasta returns a list in each element is a SeqFastadna object.
}
