% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypermutR.R
\name{remove_hypermut}
\alias{remove_hypermut}
\title{Removes hypermutation from sequence data}
\usage{
remove_hypermut(dat)
}
\arguments{
\item{dat}{The sequence data}
}
\description{
Removes hypermutation from sequence data
}
