% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypermutR.R
\name{remove_hypermut}
\alias{remove_hypermut}
\title{Removes hypermutation from sequence data}
\usage{
remove_hypermut(dat, verbose = TRUE, fix_with = FALSE,
  ancestor = "consensus", p_value = 0.05)
}
\arguments{
\item{dat}{The sequence data}

\item{verbose}{Prints name and p-value of removed/fixed sequences}

\item{fix_with}{Either false or a single letter. If not FALSE, then replace the hypermutated base with the letter indicated.}

\item{ancestor}{Either 'consensus' to indicate that the consensus sequences must be computed, or 'first' to indicate that the first sequence in the dataset should be considered to be the ancestral sequence, or the ancestral sequence itself.}

\item{p_value}{The p-value used by the one-sided fischer test.}
}
\description{
Removes hypermutation from sequence data
}
